rename rsc-renderer to rsc-handler
